+++ labb6/src/labb6/supermarketSimulator/events/CustomerEvent.java
@@ -13,7 +13,7 @@ import labb6.supermarketSimulator.ShopState;
 
 public abstract class CustomerEvent extends Event {
     protected int customerID;
-    public CustomerEvent(EventQueue eventQueue, int time, int customerID) {
+    public CustomerEvent(EventQueue eventQueue, double time, int customerID) {
         super(eventQueue, time);
         this.customerID = customerID;
     }
